refactor plugin interface and remote unity scene script
